ayurveda-chatbot/backend/requirements.txt
streamlit
requests
python-dotenv
PyPDF2
flask

